Nth root of number
Consider 3rd root of 27. Binary search because it is monotonically increasing. 
we find mid between 1 & 27 i.e., 14. If we multiply 14x14x14, it is more than 27. So, we are decreasing the search space "end" to 14. 
Mid calculated b/w 1 & 14 i.e., 7. Now 7x7x7, it is more than 27 and the same step, the size reduces to 7 and So on.
At some point we use precision as parameter used for no. of decimals we need.