Koko bananas - BS on Speed
Binary search should be applied on speed of it.
If no. of hrs. given is equals piles of length then max in the piles should be the speed
Speed calculated with 0 and max in piles, if is less than given hours we can try by reducing the mid by 1 i.e., high = mid -1