Reverse K Group
First, we have the prev pointer which is initially set to dummy. This pointer is used to keep track of the last node of the previous k-group.

Next, we have the start pointer which is set to the head of the list. This pointer is used to keep track of the first node of the current k-group.

Then we have the end pointer which is set to the same node as the start pointer. This pointer is used to keep track of the last node of the current k-group.

Next, we have the next pointer which is set to the next node after the end node. This pointer is used to keep track of the node after the current k-group.

Inside the loop, we first check if there are enough nodes left in the list to form a k-group. If not, we break out of the loop and return the dummy.next node.

If there are enough nodes left in the list, we use a for loop to reverse the current k-group. Within the for loop, we have three pointers: curr, temp, and prev.

The curr pointer is set to the start node and is used to traverse the current k-group.

The temp pointer is used to temporarily store the next node to curr. We set temp to curr.next, which is the next node in the k-group.

The prev pointer is used to keep track of the last node of the previous k-group. We set curr.next to prev, effectively reversing the direction of the link between curr and prev.

After the for loop is finished, we set the start.next node to the next node after the current k-group, which is the next pointer we defined earlier. We then update the prev pointer to be the start node of the current k-group, and update the start pointer to be the next node. We set the end pointer to the same node as the start pointer and the next pointer to the next node after the end pointer.

We continue this process until we reach the end of the list, at which point we return the dummy.next node, which is the head of the reversed list.