Maximum Points You Can Obtain from Cards
Let's assume K=3 and the maximum sum lies in the start of the array
i.e. at index 0,1,2
So, we will add the values at these 3 index to the sum.

Now, there can be a sum which could be greater than the current sum if we had selected some or all cards from the back of the array.
So, we will shift our window from front of the array to the back of the array, one element at a time.
When we shift our window, we will subtract the element we are removing from the sum and add the element we are shiting towards in our sum.

One edge case we can handle is, if the K is same the array size the answer will be always be sum of all elements in the array

Complexity
Time complexity: O(K)
Since we are only iterating through the array for K elements.

Space complexity: O(1)
No extra space is used.